+++ R/readSC.R
@@ -1,30 +1,21 @@
-#' Read single-case data from files
+#' Load single-case data from files
 #' 
-#' Use the \code{readSC} function to import single-case data from structured
-#' .csv or the \code{readSC.excel} function for importing excel files.
-#' 
+#' Use the \code{load_scdf} function to load single-case data csv, excel, or yaml files.
 #' 
 #' @aliases readSC readSC.excel
-#' @param filename A character string defining the file to be imported (e.g.
-#' \code{"SC_Anita.csv"}. If filename is left empty a dialog box for choosing
+#' @param filename A character string defining the file to be loaded (e.g.
+#' \code{"SC_Anita.csv"}. If left empty a dialog box for choosing
 #' will be opened.
-#' @param data A data frame. As an alternative to \code{filname} a dataframe could be 
-#' directly provided. 
-#' @param sep The field separator string. Values within rows have to be
-#' separated by this string. Default is \code{sep = ","}.
-#' @param dec The string used for decimal points in the file. Must be a single
-#' character. Default is \code{dec = "."}
+#' @param data A data frame. As an alternative to \code{filname}. 
 #' @param sort.labels If set TRUE, the resulting list is sorted by label names
 #' (alphabetically increasing).
-#' @param type Format of the file to be imported. Either "csv" or "excel" is
-#' possible.
-#' @param cvar Sets the variable name of the "case" variable. Deafults to \code{"case"}.
-#' @param pvar Sets the variable name of the "phase" variable. Deafults to \code{"phase"}.
-#' @param dvar Sets the variable name of the "values" variable. Deafults to \code{"values"}.
-#' @param mvar Sets the variable name of the "mt" variable. Deafults to \code{"mt"}.
-#' @param phase.names A character vector with phase names. Deafults to the phase names provided 
-#' in the phase variable.
-#' @param \dots Further arguments passed to the \code{\link{read.table}}
+#' @param type Format of the file to be loaded. Either "csv", "xlsx", "xls", "excel", "yml" is possible. By deafult (NA) the type is extracted from the file extension.
+#' @param cvar Sets the variable name of the "case" variable. Defaults to \code{"case"}.
+#' @param pvar Sets the variable name of the "phase" variable. Defaults to \code{"phase"}.
+#' @param dvar Sets the variable name of the "values" variable. Defaults to \code{"values"}.
+#' @param mvar Sets the variable name of the "mt" variable. Defaults to \code{"mt"}.
+#' @param phase.names A character vector with phase names. Defaults to the phase names provided in the phase variable.
+#' @param \dots Further arguments passed to the repective read function (\code{\link{read.table}}, \code{\link{read_excel}}, \code{\link{read_yaml}}
 #' command.
 #' @return Returns a single-case data frame. See \code{\link{scdf}} to learn
 #' about the format of these data frames.
@@ -34,47 +25,61 @@
 #' @examples
 #' 
 #' ## Read SC-data from a file named "study1.csv" in your working directory
-#' # study1 <- readSC("study1.csv")
+#' # study1 <- load_scdf("study1.csv")
 #' 
 #' ## Read SC-data from a .csv-file with semicolon as field and comma as decimal separator
-#' # study2 <- readSC("study2.csv", sep = ";", dec = ",")
+#' # study2 <- load_scdf("study2.csv", sep = ";", dec = ",")
 #' 
 #' ## writeSc and readSC
-#' filename <- file.path(tempdir(),"test.csv")
+#' filename <- file.path(tempdir(), "test.csv")
 #' writeSC(exampleA1B1A2B2_zvt, filename)
-#' dat <- readSC(filename, cvar = "case", pvar = "part", dvar = "zvt", mvar = "day")
+#' dat <- load_scdf(filename, cvar = "case", pvar = "part", dvar = "zvt", mvar = "day")
 #' res1 <- describeSC(exampleA1B1A2B2_zvt)$descriptives
 #' res2 <- describeSC(dat)$descriptives
-#' identical(res1,res2)
+#' all.equal(res1,res2)
 #' 
 #' @export
-readSC <- function(filename = NULL, data = NULL, sep = ",", dec = ".", sort.labels = FALSE, cvar = "case", pvar = "phase", dvar = "values", mvar = "mt", phase.names = NULL, type = "csv", ...) {
-  if (is.null(filename) && is.null(data)) {
+load_scdf <- function(filename, 
+                      data = NULL, 
+                      sort.labels = FALSE, 
+                      cvar = "case", 
+                      pvar = "phase", 
+                      dvar = "values", 
+                      mvar = "mt", 
+                      phase.names = NULL, 
+                      type = NA, 
+                      ...) {
+  
+  if (missing(filename) && is.null(data)) {
     filename <- file.choose()
-    cat("Import file", filename, "\n\n")
+    cat("Load file", filename, "\n\n")
   }
+  
+  if (is.na(type)) {
+    type <- substring(filename, regexpr("\\.([[:alnum:]]+)$", filename) + 1)
+  }
+  
   if (!is.null(data)) {
     type <- "data"
     dat <- as.data.frame(data)
   }
   
-  if (type == "csv")
-    dat <- utils::read.table(filename, header = TRUE, sep = sep, dec = dec, stringsAsFactors = FALSE,...)
-  if (type == "excel") {
-    if (!requireNamespace("readxl", quietly = TRUE)) {
-      stop("Package readxl needed for this function to work. Please install it.",
-           call. = FALSE)
-    }
-    #stop("Excel import currently not supported.")
+  if (type == "csv") {
+    dat <- utils::read.table(
+      filename, 
+      header = TRUE,
+      stringsAsFactors = FALSE,
+      ...
+    )
+  }
+  
+  if (type %in% c("yml", "yaml")) return(.load_yml(filename, ...))
+  
+  if (type %in% c("excel", "xlsx", "xls")) {
     dat <- as.data.frame(readxl::read_excel(filename, ...))
   }
 
-  VARS <- c(cvar, pvar, dvar, mvar)
   columns <- ncol(dat)
- 
-  #pos <- match(VARS, names(dat))
-  #pos.rest <- which(!(1:columns %in% pos))
-  #dat <- dat[, c(pos, pos.rest)]
 
   if (!sort.labels) {
     dat[[cvar]] <- factor(dat[[cvar]], levels = unique(dat[[cvar]]))
@@ -91,21 +96,83 @@ readSC <- function(filename = NULL, data = NULL, sep = ",", dec = ".", sort.labe
   dat <- lapply(dat, function(x) x[, 2:columns])
   for(i in 1:length(dat)) row.names(dat[[i]]) <- 1:nrow(dat[[i]])
   names(dat) <- lab
-  cat("Imported", length(dat), "cases.\n")
+  cat("Loaded", length(dat), "cases.\n")
 
-  class(dat) <- c("scdf","list")
+  class(dat) <- c("scdf", "list")
   scdf_attr(dat, .opt$phase) <- pvar
   scdf_attr(dat, .opt$dv)    <- dvar
   scdf_attr(dat, .opt$mt)    <- mvar
   
   
-  return(dat)
+  dat
 }
 
-#' @rdname readSC
+#' @rdname load_scdf
 #' @export
 readSC.excel <- function(...) {
-  readSC(..., type = "excel")
+  
+  load_scdf(..., type = "excel")
   
 }
 
+#' @rdname load_scdf
+#' @export
+readSC <- function(...) {
+  load_scdf(...)
+}
+
+
+#' @export
+.load_yml <- function(filename, ...) {
+  
+  out <- yaml::read_yaml(filename, ...)
+  
+  case_names <- names(out)
+  
+  extract <- function(x) {
+    
+    if (is.null(x[["dvar"]])) x[["dvar"]] <- "values"
+    if (is.null(x[["mvar"]])) x[["mvar"]] <- "mt"
+    if (is.null(x[["pvar"]])) x[["pvar"]] <- "phase"
+    
+    dvar <- x[["dvar"]]
+    pvar <- x[["pvar"]]
+    mvar <- x[["mvar"]]
+    
+    x[["dvar"]] <- NULL
+    x[["mvar"]] <- NULL
+    x[["pvar"]] <- NULL
+    
+    phase <- numeric(0)
+    values <- numeric(0)
+    
+    for(i in seq_along(x[[dvar]])) {
+      phase <- c(
+        phase, rep(names(x[[dvar]][i]), length(x[[dvar]][[i]])))
+      values <- c(values, x[[dvar]][[i]])
+    }
+    
+    x[[pvar]] <- phase
+    x[[dvar]] <- values
+    
+    if (is.null(x[[mvar]])) x[[mvar]] <- 1:length(x[[dvar]])
+    
+    x <- list(as.data.frame(x))
+    scdf_attr(x, "var.phase") <- pvar
+    scdf_attr(x, "var.values") <- dvar
+    scdf_attr(x, "var.mt") <- mvar
+    class(x) <- c("scdf", "list")
+    
+    x
+    
+  }
+  
+  out <- lapply(out, extract)
+  
+  for(i in 1:length(out)) names(out[[i]]) <- case_names[i]
+  
+  for(i in 2:length(out)) scdf <- c(out[[i-1]], out[[i]])
+  
+  
+  scdf
+}